simplify components
